memul: Add a flag to disable memcpy_base opt
The memcpy_base opt is not relevant to 8226.
To avoid degradation for non-krait cores, add a
flag to control the necessity of optimization.

If the flag is set to true, the opt won't be used.
Instead, the memcpy/memcpy_base are used from
cortex-a15. Testing shows that the Quadrant memory
scores has improved and comparable to KK.

CRs-fixed: 790279

Change-Id: I59169166fddd3a6e0744644233ff275a9639e1d3